+++ docs/.vuepress/config.js
@@ -2,25 +2,51 @@ module.exports = {
   locales: {
     '/': {
       lang: 'en-US',
-      title: 'Vue CLI 3',
+      title: 'Vue CLI',
       description: '🛠️ Standard Tooling for Vue.js Development'
     },
     '/zh/': {
       lang: 'zh-CN',
-      title: 'Vue CLI 3',
+      title: 'Vue CLI',
       description: '🛠️ Vue.js 开发的标准工具'
     },
     '/ru/': {
       lang: 'ru',
-      title: 'Vue CLI 3',
+      title: 'Vue CLI',
       description: '🛠️ Стандартный инструментарий для разработки на Vue.js'
     }
   },
-  serviceWorker: true,
   head: [
     ['link', { rel: 'icon', href: '/favicon.png' }],
+    ['link', { rel: 'manifest', href: '/manifest.json' }],
+    ['meta', { name: 'theme-color', content: '#3eaf7c' }],
+    ['meta', { name: 'apple-mobile-web-app-capable', content: 'yes' }],
+    ['meta', { name: 'apple-mobile-web-app-status-bar-style', content: 'black' }],
+    ['link', { rel: 'apple-touch-icon', href: `/icons/apple-touch-icon-152x152.png` }],
+    ['link', { rel: 'mask-icon', href: '/icons/safari-pinned-tab.svg', color: '#3eaf7c' }],
+    ['meta', { name: 'msapplication-TileImage', content: '/icons/msapplication-icon-144x144.png' }],
+    ['meta', { name: 'msapplication-TileColor', content: '#000000' }]
   ],
-  theme: 'vue',
+  plugins: {
+    '@vuepress/pwa': {
+      serviceWorker: true,
+      updatePopup: {
+        '/': {
+          message: "New content is available.",
+          buttonText: "Refresh"
+        },
+        '/zh/': {
+          message: "发现新内容可用",
+          buttonText: "刷新"
+        },
+        '/ru/': {
+          message: 'Доступно обновление контента',
+          buttonText: 'Обновить'
+        }
+      }
+    }
+  },
+  theme: '@vuepress/theme-vue',
   themeConfig: {
     repo: 'vuejs/vue-cli',
     docsDir: 'docs',
@@ -37,12 +63,6 @@ module.exports = {
         selectText: 'Languages',
         lastUpdated: 'Last Updated',
         editLinkText: 'Edit this page on GitHub',
-        serviceWorker: {
-          updatePopup: {
-            message: "New content is available.",
-            buttonText: "Refresh"
-          }
-        },
         nav: [
           {
             text: 'Guide',
@@ -103,7 +123,8 @@ module.exports = {
                 '/guide/webpack',
                 '/guide/mode-and-env',
                 '/guide/build-targets',
-                '/guide/deployment'
+                '/guide/deployment',
+                '/guide/troubleshooting'
               ]
             }
           ],
@@ -134,12 +155,6 @@ module.exports = {
         selectText: '选择语言',
         lastUpdated: '上次编辑时间',
         editLinkText: '在 GitHub 上编辑此页',
-        serviceWorker: {
-          updatePopup: {
-            message: "发现新内容可用",
-            buttonText: "刷新"
-          }
-        },
         nav: [
           {
             text: '指南',
@@ -223,12 +238,6 @@ module.exports = {
         selectText: 'Переводы',
         lastUpdated: 'Последнее обновление',
         editLinkText: 'Изменить эту страницу на GitHub',
-        serviceWorker: {
-          updatePopup: {
-            message: 'Доступно обновление контента',
-            buttonText: 'Обновить'
-          }
-        },
         nav: [
           {
             text: 'Руководство',