+++ packages/@vue/cli-plugin-babel/__tests__/transpileDependencies.spec.js
@@ -40,6 +40,7 @@ beforeAll(async () => {
 
   $packageJson = JSON.parse($packageJson)
   $packageJson.dependencies['external-dep'] = '1.0.0'
+  $packageJson.dependencies['@scope/external-dep'] = '1.0.0'
   $packageJson = JSON.stringify($packageJson)
 
   await project.write(
@@ -70,10 +71,23 @@ test('dep from node_modules should not been transpiled', async () => {
 test('dep from node_modules should been transpiled', async () => {
   await project.write(
     'vue.config.js',
-    `module.exports = { transpileDependencies: ['external-dep'] }`
+    `module.exports = { transpileDependencies: ['external-dep', '@scope/external-dep'] }`
   )
   await project.run('vue-cli-service build')
   expect(await readVendorFile()).toMatch('return "__TEST__"')
 
   expect(await readVendorFile()).toMatch('return "__SCOPE_TEST__"')
+})
+
+// https://github.com/vuejs/vue-cli/issues/3057
+test('only transpile package with same name specified in transpileDependencies', async () => {
+  await project.write(
+    'vue.config.js',
+    `module.exports = { transpileDependencies: ['babel-transpile-deps'] }`
+  )
+  try {
+    await project.run('vue-cli-service build')
+  } catch (e) {}
+  expect(await readVendorFile()).toMatch('() => "__TEST__"')
+  expect(await readVendorFile()).toMatch('() => "__SCOPE_TEST__"')
 })