+++ packages/@vue/cli-plugin-typescript/generator/index.js
@@ -9,7 +9,7 @@ module.exports = (api, {
 
   api.extendPackage({
     devDependencies: {
-      typescript: api.hasPlugin('eslint') ? '~3.1.1' : '^3.0.0'
+      typescript: '^3.2.1'
     }
   })
 
@@ -17,7 +17,7 @@ module.exports = (api, {
     api.extendPackage({
       dependencies: {
         'vue-class-component': '^6.0.0',
-        'vue-property-decorator': '^7.0.0'
+        'vue-property-decorator': '^8.0.0'
       }
     })
   }