+++ packages/@vue/cli-plugin-typescript/lib/tslint.js
@@ -29,6 +29,7 @@ module.exports = function lint (args = {}, api, silent) {
       if (isVueFile(file)) {
         const parts = vueFileCache.get(path.normalize(file))
         if (parts) {
+          parts.content = content
           const { before, after } = parts
           content = `${before}\n${content.trim()}\n${after}`
         }
@@ -42,14 +43,20 @@ module.exports = function lint (args = {}, api, silent) {
   }
 
   const parseTSFromVueFile = file => {
+    // If the file has already been cached, don't read the file again. Use the cache instead.
+    if (vueFileCache.has(file)) {
+      return vueFileCache.get(file)
+    }
+
     const content = fs.readFileSync(file, 'utf-8')
     const { script } = vueCompiler.parseComponent(content, { pad: 'line' })
     if (script && /^tsx?$/.test(script.lang)) {
       vueFileCache.set(file, {
         before: content.slice(0, script.start),
-        after: content.slice(script.end)
+        after: content.slice(script.end),
+        content: script.content
       })
-      return script.content
+      return script
     }
   }
 
@@ -61,8 +68,9 @@ module.exports = function lint (args = {}, api, silent) {
     const getSourceFile = program.getSourceFile
     program.getSourceFile = function (file, languageVersion, onError) {
       if (isVueFile(file)) {
-        const script = parseTSFromVueFile(file) || ''
-        return ts.createSourceFile(file, script, languageVersion, true)
+        const { content, lang = 'js' } = parseTSFromVueFile(file) || { content: '', lang: 'js' }
+        const contentLang = ts.ScriptKind[lang.toUpperCase()]
+        return ts.createSourceFile(file, content, languageVersion, true, contentLang)
       } else {
         return getSourceFile.call(this, file, languageVersion, onError)
       }