+++ packages/@vue/cli-service/lib/commands/serve.js
@@ -34,7 +34,6 @@ module.exports = (api, options) => {
     const isProduction = process.env.NODE_ENV === 'production'
 
     const url = require('url')
-    const path = require('path')
     const chalk = require('chalk')
     const webpack = require('webpack')
     const WebpackDevServer = require('webpack-dev-server')
@@ -140,9 +139,7 @@ module.exports = (api, options) => {
       clientLogLevel: 'none',
       historyApiFallback: {
         disableDotRule: true,
-        rewrites: [
-          { from: /./, to: path.posix.join(options.publicPath, 'index.html') }
-        ]
+        rewrites: genHistoryApiFallbackRewrites(options.publicPath, options.pages)
       },
       contentBase: api.resolve('public'),
       watchContentBase: !isProduction,
@@ -201,8 +198,12 @@ module.exports = (api, options) => {
 
         let copied = ''
         if (isFirstCompile && args.copy) {
-          require('clipboardy').write(urls.localUrlForBrowser)
-          copied = chalk.dim('(copied to clipboard)')
+          try {
+            require('clipboardy').writeSync(urls.localUrlForBrowser)
+            copied = chalk.dim('(copied to clipboard)')
+          } catch (_) {
+            /* catch exception if copy to clipboard isn't supported (e.g. WSL), see issue #3476 */
+          }
         }
 
         const networkUrl = publicUrl
@@ -301,6 +302,23 @@ function checkInContainer () {
     const content = fs.readFileSync(`/proc/1/cgroup`, 'utf-8')
     return /:\/(lxc|docker|kubepods)\//.test(content)
   }
+}
+
+function genHistoryApiFallbackRewrites (baseUrl, pages = {}) {
+  const path = require('path')
+  const multiPageRewrites = Object
+    .keys(pages)
+    // sort by length in reversed order to avoid overrides
+    // eg. 'page11' should appear in front of 'page1'
+    .sort((a, b) => b.length - a.length)
+    .map(name => ({
+      from: new RegExp(`^/${name}`),
+      to: path.posix.join(baseUrl, pages[name].filename || `${name}.html`)
+    }))
+  return [
+    ...multiPageRewrites,
+    { from: /./, to: path.posix.join(baseUrl, 'index.html') }
+  ]
 }
 
 module.exports.defaultModes = {