+++ packages/@vue/cli-service/types/ProjectOptions.d.ts
@@ -1,0 +1,60 @@
+import ChainableWebpackConfig from 'webpack-chain'
+import { WebpackOptions } from 'webpack/declarations/WebpackOptions'
+
+type PageEntry = string;
+
+interface PageConfig {
+  entry: PageEntry;
+  [key: string]: any;
+}
+
+interface LoaderOptions {
+  css?: object;
+  sass?: object;
+  less?: object;
+  stylus?: object;
+  postcss?: object;
+}
+
+// mini-css-extract-plugin options
+interface ExtractOptions {
+  filename?: string;
+  chunkFilename?: string;
+}
+
+interface CSSOptions {
+  modules?: boolean;
+  extract?: boolean | ExtractOptions;
+  sourceMap?: boolean;
+  loaderOptions?: LoaderOptions;
+}
+
+export interface ProjectOptions {
+  publicPath?: string;
+  outputDir?: string;
+  assetsDir?: string;
+  indexPath?: string;
+  filenameHashing?: boolean;
+  runtimeCompiler?: boolean;
+  transpileDependencies?: Array<string | RegExp>;
+  productionSourceMap?: boolean;
+  parallel?: boolean;
+  devServer?: object;
+  pages?: {
+    [key: string]: PageEntry | PageConfig;
+  };
+  crossorigin?: '' | 'anonymous' | 'use-credentials';
+  integrity?: boolean;
+
+  css?: CSSOptions;
+
+  chainWebpack?: (config: ChainableWebpackConfig) => void;
+  configureWebpack?: WebpackOptions | ((config: WebpackOptions) => (WebpackOptions | void));
+
+  lintOnSave?: boolean | 'error';
+  pwa?: object;
+
+  pluginOptions?: object;
+}
+
+export type ConfigFunction = () => ProjectOptions