+++ packages/@vue/cli-service/types/index.d.ts
@@ -1,0 +1,1 @@
+export { ProjectOptions, ConfigFunction } from './ProjectOptions'