+++ packages/@vue/cli-ui/src/components/prompt/PromptsList.vue
@@ -79,7 +79,7 @@ export default {
 .group-name
   padding $padding-item $padding-item ($padding-item / 2)
   font-size 1.6em
-  font-weight lighter
+  font-weight 300
   color $vue-ui-color-accent
   .vue-ui-dark-mode &
     color lighten($vue-ui-color-accent, 60%)