+++ packages/@vue/cli-ui/src/i18n.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import VueI18n from 'vue-i18n'
 import deepmerge from 'deepmerge'
-import VueTimeago from 'vue-timeago'
+import VueTimeago, { createTimeago } from 'vue-timeago'
 
 Vue.use(VueI18n)
 
@@ -50,11 +50,11 @@ async function autoDetect () {
     }
 
     const dateFnsLocale = i18n.locale.toLowerCase().replace(/-/g, '_')
-    Vue.component('VueTimeago', VueTimeago.createTimeago({
+    Vue.component('VueTimeago', createTimeago({
       name: 'VueTimeago',
       locale: i18n.locale,
       locales: {
-        [i18n.locale]: require(`date-fns/locale/${dateFnsLocale}`)
+        [i18n.locale]: require(`date-fns/locale/${dateFnsLocale}/index.js`)
       }
     }))
   }